# vigil-requirements/OPENAPI_SPECIFICATION.yaml
openapi: 3.1.0
info:
  title: Vigil Dashboard API
  description: |
    Vigil Dashboard API - A security-focused home network monitoring system.
    
    ## Authentication
    This API uses Bearer token authentication for all endpoints. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Security Considerations
    - All requests must include proper CSRF tokens for state-changing operations
    - Rate limiting is enforced: 100 requests per minute per IP
    - Sensitive data is encrypted at rest and in transit
    
  version: 2.0.0
  contact:
    name: Vigil Security Team
    email: security@vigil.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v2
    description: Local development server
  - url: https://api.vigil.local/api/v2
    description: Production server
security:
  - BearerAuth: []
  - CsrfToken: []
...
paths:
...
  /devices:
    get:
      tags:
        - Devices
      summary: List devices
      description: Get all network devices with optional filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [observing, blocked, trusted, quarantined]
          description: Filter by containment status
        - name: search
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: Search by MAC, IP, hostname, or nickname
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /devices/{deviceId}:
    get:
      tags:
        - Devices
      summary: Get device
      description: Get detailed information about a specific device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Device ID
      responses:
        "200":
          description: Device details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - Devices
      summary: Update device
      description: Update device nickname, hostname, or type
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdateRequest"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceUpdateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token for state-changing operations
  schemas:
...
    DeviceResponse:
      type: object
      required:
        - id
        - mac
        - ip
        - trust_score
        - containment_status
        - online
        - last_seen
        - first_seen
      properties:
        id:
          type: integer
          minimum: 1
        mac:
          type: string
          pattern: "^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$"
          example: "AA:BB:CC:DD:EE:FF"
        ip:
          type: string
          format: ipv4
          example: "192.168.1.100"
        hostname:
          type: string
          nullable: true
          example: "iPhone-John"
        nickname:
          type: string
          nullable: true
          example: "John's Phone"
        vendor:
          type: string
          nullable: true
          example: "Apple Inc."
        device_type:
          type: string
          nullable: true
          enum: [phone, laptop, desktop, tablet, iot, tv, speaker, unknown, agent]
        trust_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 75.5
        containment_status:
          type: string
          enum: [observing, blocked, trusted, quarantined]
          example: observing
        online:
          type: boolean
        last_seen:
          type: string
          format: date-time
        first_seen:
          type: string
          format: date-time
    DeviceListResponse:
      type: object
      required:
        - devices
        - total_count
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceResponse"
        total_count:
          type: integer
          minimum: 0
    DeviceUpdateRequest:
      type: object
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 100
          example: "John's iPhone"
        hostname:
          type: string
          minLength: 1
          maxLength: 100
        device_type:
          type: string
          enum: [phone, laptop, desktop, tablet, iot, tv, speaker, unknown]
    DeviceUpdateResponse:
      type: object
      required:
        - success
        - message
        - device
      properties:
        success:
          type: boolean
        message:
          type: string
          example: "Device updated successfully"
        device:
          $ref: "#/components/schemas/DeviceResponse"
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
        error_code:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Invalid request parameters"
            error_code: BAD_REQUEST
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Invalid or expired token"
            error_code: UNAUTHORIZED
            timestamp: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Resource not found"
            error_code: NOT_FOUND
            timestamp: "2024-01-15T10:30:00Z"
...